Add a new mechanism for tracking collections; also, remove ChildOf.
The ChildOf component is not helpful in practice - it's better
to have a variety of different components that can express
collection-like relationships between entities.

Now, any component with a `collectionEntity: Entity` property
will implicitly be tracked by the world as a many-to-one
relationship, allowing to obtain the set of many entities related
to a given one by using the new `world.getCollected` function.

The intent is that this will be used for many different
kinds of relationships, each with their own separate semantics
and related systems.

The first of those is `PositionWithin`, which implies that the
`Position` of the collected entity is to be transformed during
drawing such that the `Position` of the collection entity
is its origin. Note that this means the `Position` of the
collected entity is now expressing relative rather than absolute
position.